feat(layers): expand-in-place card, no portal/backdrop (matches legend)
LayerSheet was a fullscreen-overlay modal with mobile backdrop blur.
The Map legend's collapse/expand-in-place pattern reads cleaner — the
user keeps the map visible while toggling, and there's no separate
trigger element that disappears when the sheet opens. Port the same
behavior to Layers.

Single component (`<LayerSheet />`) replaces the prior trigger-+-modal
split:
- Collapsed: round h-11 w-11 icon button at the previous trigger
  position. Same active-dot indicator (any toggle off its default).
- Expanded: card grows upward from the bottom-anchored wrapper to
  fit the tab bar + active-tab content. Header shows the Layers icon
  + 'Layers' label + X close button.
- No createPortal, no fixed-inset overlay, no mobile backdrop.
- Width transitions w-11 → w-80 (or w-[calc(100%-1.5rem)] on mobile);
  the wrapper's bottom anchor stays at the same calc as the legend's
  so the two cards line up at the same baseline.

`<LayerSheetTrigger />` deleted; `__root.tsx` updated to mount only
`<LayerSheet />`. The data-mapky-layer-trigger attribute moves to the
new wrapper so the legend's outside-click logic (in MapLegends header
icon switching) keeps targeting the same element.
