+++ src/routes/__root.tsx
@@ -12,7 +12,6 @@ import { BtcOverlayLayer } from "@/components/map/BtcOverlayLayer";
 import { MapLegends } from "@/components/map/MapLegends";
 import { Buildings3DLayer } from "@/components/map/Buildings3DLayer";
 import { LayerSheet } from "@/components/map/LayerSheet";
-import { LayerSheetTrigger } from "@/components/map/LayerSheetTrigger";
 import { IconRail } from "@/components/sidebar/IconRail";
 import { MobileNavDrawer } from "@/components/sidebar/MobileNavDrawer";
 import { SearchBar } from "@/components/sidebar/SearchBar";
@@ -48,7 +47,6 @@ function RootLayout() {
           <DirectionsLayer />
           <CaptureCreationPanel />
           <Outlet />
-          <LayerSheetTrigger />
           <LayerSheet />
           <MapLegends />
         </div>
